Ouverture modale édition de l'arrêté de paiement + création du document
