+++ .docker/nginx/conf/mon-indemnisation.justice.gouv.dev.conf
@@ -14,6 +14,8 @@ server {
     error_log /var/log/nginx/mon-indemnisation.justice.gouv.dev.error.log;
     access_log /var/log/nginx/mon-indemnisation.justice.gouv.dev.access.log;
 
+    client_max_body_size 10M;
+
     location ~ /preview {
         try_files $uri @vite;
     }
