+++ .docker/nginx/conf/mon-indemnisation.justice.gouv.test.conf
@@ -14,6 +14,8 @@ server {
     error_log /var/log/nginx/mon-indemnisation.justice.gouv.test.error.log;
     access_log /var/log/nginx/mon-indemnisation.justice.gouv.test.access.log;
 
+    client_max_body_size 10M;
+
     location / {
         try_files $uri @symfony;
     }
