+++ assets/apps/agent/dossiers/components/consultation/index.ts
@@ -1,11 +1,13 @@
 import { AttributionDossier } from "./AttributionDossier";
 import { ClotureDossier } from "./ClotureDossier";
 import { DecisionDossier } from "./DecisionDossier";
-import { ValidationDossier } from "./ValidationDossier";
+import { ValidationDecisionDossier } from "./ValidationDecisionDossier";
+import { ValidationAcceptationDossier } from "./ValidationAcceptationDossier";
 
 export {
   AttributionDossier,
   ClotureDossier,
   DecisionDossier,
-  ValidationDossier,
+  ValidationDecisionDossier,
+  ValidationAcceptationDossier,
 };
