+++ config/packages/web_profiler.yaml
@@ -1,6 +1,6 @@
 when@dev:
     web_profiler:
-        toolbar: true
+        toolbar: false
         intercept_redirects: false
 
     framework:
