+++ templates/courrier/_pied_de_page.html.twig
@@ -1,0 +1,7 @@
+<footer>
+    <p>
+        13, place Vendôme – 75042 Paris Cedex 01<br/>
+        Téléphone : 01 44 77 60 60<br/>
+        www.justice.gouv.fr
+    </p>
+</footer>
