+++ templates/email/agent_indemnisation_approuvee.twig
@@ -17,7 +17,7 @@
     <br/>
     <br/>
     <p>
-        {% if dossier.requerant.estFeminin() %}La requérante{% else %}Le requérant{% endif %} {{ dossier.requerant.nomComplet }}
+        {% if dossier.requerant.estFeminin() %}La requérante{% else %}Le requérant{% endif %} {{ dossier.requerant.nomCourant }}
         vient d'approuver la proposition d'indemnisation qui lui est
         faite sur le dossier de référence {{ dossier.reference }}.
     </p>
